update add_to_wishlist view and remove arrow from toast messages

+++ wishlist/views.py
@@ -22,26 +22,28 @@ def wishlist(request):
         request, 'wishlist/wishlist.html', {'user_wishlist': user_wishlist})
 
 
-@login_required
 def add_to_wishlist(request, product_id):
     '''
     A View to add/remove item from the wishlist
     '''
 
     product = get_object_or_404(Product, pk=product_id)
-    user_profile = UserProfile.objects.get(user=request.user)
 
-    
-    if Wishlist.objects.filter(user_profile=user_profile,product=product).exists():
-        Wishlist.objects.get(user_profile=user_profile,product=product).delete()
-        messages.success(
-            request, f'{product.name} has been removed from your Wishlist.')
+    if request.user.is_authenticated:
+        user_profile = UserProfile.objects.get(user=request.user)
+        if Wishlist.objects.filter(user_profile=user_profile,product=product).exists():
+            Wishlist.objects.get(user_profile=user_profile,product=product).delete()
+            messages.success(
+                request, f'{product.name} has been removed from your Wishlist.')
+        else:
+            wishlist_item = Wishlist.objects.create(
+                user_profile=user_profile, product=product)
+            messages.success(
+                request,
+                f'{wishlist_item.product.name} added to Wishlist successfully!'
+            )
+        return HttpResponseRedirect(request.META["HTTP_REFERER"])
     else:
-        wishlist_item = Wishlist.objects.create(
-            user_profile=user_profile, product=product)
-        messages.success(
-            request,
-            f'{wishlist_item.product.name} added to Wishlist successfully!'
-        )
-    return HttpResponseRedirect(request.META["HTTP_REFERER"])
+        messages.error(request, 'You need to log in to add item to your Wishlist')
+        return HttpResponseRedirect(request.META["HTTP_REFERER"])
